requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
filelock==3.18.0
fsspec==2025.3.2
huggingface-hub==0.30.2
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
PyJWT==2.9.0
pyotp==2.9.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rest-framework-simplejwt==0.0.2
safetensors==0.5.3
setuptools==78.1.0
soupsieve==2.6
sqlparse==0.5.3
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.13.1
uritemplate==4.1.1
urllib3==2.3.0
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.3.2
transformers==4.37.2
